refactor a controller/route and update imports for CustomError
