+++ src/controllers/channelController/index.ts
@@ -1,0 +1,11 @@
+import { getAllChannels } from './getAllChannels';
+import { createChannel } from './createChannel';
+import { addUserToChannel } from './addUserToChannel';
+
+const channelController = {
+  getAllChannels,
+  createChannel,
+  addUserToChannel
+}
+
+export default channelController;
