+++ src/controllers/index.ts
@@ -1,0 +1,1 @@
+export { default as channelController } from './channelController';
