+++ src/controllers/usersController.ts
@@ -1,13 +1,13 @@
 import { Request, Response } from "express";
 import { User } from "../models/user";
-import { customError } from "../middlewares/globalErrorMiddleware";
+import { CustomError } from '../middlewares/middleware';
 
 export const createNewUser = async (req: Request, res: Response) => {
   const { userName, userEmail, firstName, lastName } = req.body;
 
   //   TODO: validate the input data with an efficient way
   if (!userName || !userEmail || !firstName || !lastName) {
-    throw new customError("Please provide all the details", 400);
+    throw new CustomError("Please provide all the details", 400);
   }
   // check if user already exists
   const user = await User.findOne({
@@ -16,7 +16,7 @@ export const createNewUser = async (req: Request, res: Response) => {
     },
   });
   if (user) {
-    throw new customError("User already exists", 400);
+    throw new CustomError("User already exists", 400);
   }
 
   const user1 = await User.create({
